requirements.txt
numpy
tensorflow==2.3.0
pyyaml
tensorflow_datasets